requirements.txt
discord.py==1.3.4
aiohttp==3.6.2
aiodns==2.0.0
urlextract==1.0.0
